[Enhancement] [RS232] [allowing 115200 baudrate] [Cleanup] #278 [RS232] [Added RS232 channel number. Adapted conf_template]
Reviewed by Martin Wodok.

+++ library/xgpl_src/IsoAgLib/hal/pc/hal_simulator/hal_simulator_c.h
@@ -79,7 +79,7 @@ class HALSimulator_c
 			{ return 0; }
 
 	// RS-232
-		virtual void init_rs232( uint16_t baudrate,uint8_t bMode,uint8_t bStoppbits,bool bitSoftwarehandshake, uint8_t comport )
+		virtual void init_rs232( uint32_t baudrate,uint8_t bMode,uint8_t bStoppbits,bool bitSoftwarehandshake, uint8_t comport )
 			{}
 		virtual void write_rs232( uint8_t aui8_channel, const uint8_t * lpBuf, uint16_t dwToWrite )
 			{}
